2011-10-18  Andrew Stubbs  <ams@codesourcery.com>
	PR tree-optimization/50717

	gcc/
	* tree-ssa-math-opts.c (is_widening_mult_p): Remove the 'type'
	parameter.  Calculate 'type' from stmt.
	(convert_mult_to_widen): Update call the is_widening_mult_p.
	(convert_plusminus_to_widen): Likewise.

	gcc/testsuite/
	* gcc.dg/pr50717-1.c: New file.
	* gcc.target/arm/wmul-12.c: Correct types.
	* gcc.target/arm/wmul-8.c: Correct types.




git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@180164 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.target/arm/wmul-12.c
@@ -5,8 +5,8 @@
 long long
 foo (int *b, int *c)
 {
-  int tmp = *b * *c;
-  return 10 + (long long)tmp;
+  long long tmp = (long long)*b * *c;
+  return 10 + tmp;
 }
 
 /* { dg-final { scan-assembler "smlal" } } */
